+++ webpack.config.js
@@ -3,14 +3,22 @@
 const webpack = require("webpack");
 const path = require("path");
 const PROD = (process.env.NODE_ENV === 'production');
+const packageData = require("./package.json");
+
+let plugins = [
+  new webpack.DefinePlugin({
+    __VERSION__: JSON.stringify(packageData.version),
+    __NAME__: JSON.stringify(packageData.name)
+  })
+];
+
+if (PROD) {
+  plugins.push(new webpack.optimize.UglifyJsPlugin({sourceMap: true}));
+}
 
 module.exports = {
   context: __dirname + "/src",
-  entry: PROD ? {
-    "ovpProvider.min": "k-provider/ovp/ovp-provider.js",
-    "ottProvider.min": "k-provider/ott/ott-provider.js",
-    "statsService.min": "k-provider/ovp/services/stats-service.js"
-  } : {
+  entry: {
     "ovpProvider": "k-provider/ovp/ovp-provider.js",
     "ottProvider": "k-provider/ott/ott-provider.js",
     "statsService": "k-provider/ovp/services/stats-service.js"
@@ -23,7 +31,7 @@ module.exports = {
     devtoolModuleFilenameTemplate: "./providers/[resource-path]",
   },
   devtool: 'source-map',
-  plugins: PROD ? [new webpack.optimize.UglifyJsPlugin({sourceMap: true})] : [],
+  plugins: plugins,
   module: {
     rules: [{
       test: /\.js$/,